refactor: distributes actions
