+++ src/store/reducers/user/action.js
@@ -1,0 +1,12 @@
+import {createAction} from '@reduxjs/toolkit';
+import {ActionType} from '../../../consts';
+
+const requireAuthorization = createAction(ActionType.REQUIRED_AUTHORIZATION, (status) => {
+  return {
+    payload: status,
+  };
+});
+
+export {
+  requireAuthorization,
+};
